refactor(livekit-cf): restructure client and sync components
- Move client components to dedicated directory structure
- Move sync backend to dedicated directory
- Consolidate exports in do/index.ts
- Update worker.ts to directly use Hono router instead of separate router module
- Remove obsolete router module files

This restructuring improves the organization of the livekit-cf package by grouping related components together and removing unnecessary indirection.
