--- packages/livekit-cf/src/router/index.ts
@@ -1,3 +1,0 @@
-import { app } from "./app";
-
-export const fetch = app.fetch;
